refactor(runtime-vapor): convert DynamicFragment.hydrate to a free function
Move the hydration anchor-resolution logic out of the DynamicFragment prototype into hydrateDynamicFragmentAnchor() in dom/hydrateFragment.ts, along with its helpers (CloseAnchorOwner, getDynamicCloseOwner, queueAnchorInsert, isReusableDynamicFragmentAnchor) and the deferred hydration cursor handling (prepareDeferredHydrationAnchor).

fragment.ts now only value-imports the two entry points; the reverse dependency from dom/hydrateFragment.ts stays type-only. Behavior is unchanged and the optional interop/teleport `hydrate` field on VaporFragment is untouched.
