+++ packages/runtime-vapor/src/fragment.ts
@@ -1,16 +1,11 @@
 import { EffectScope, type ShallowRef, setActiveSub } from '@vue/reactivity'
-import {
-  createComment,
-  createTextNode,
-  parentNode as getParentNode,
-} from './dom/node'
+import { createComment, createTextNode } from './dom/node'
 import {
   type Block,
   type BlockFn,
   EMPTY_BLOCK,
   type TransitionOptions,
   type VaporTransitionHooks,
-  findBlockBoundary,
   insert,
   isValidBlock,
   remove,
@@ -20,26 +15,11 @@ import {
   type TransitionHooks,
   type VNode,
   currentInstance,
-  queuePostFlushCb,
   setCurrentInstance,
 } from '@vue/runtime-dom'
 import type { VaporComponentInstance } from './component'
 import type { NodeRef } from './apiTemplateRef'
-import {
-  advanceHydrationNode,
-  cleanupHydrationTail,
-  currentHydrationNode,
-  enterHydrationBoundary,
-  isComment,
-  isHydrating,
-  isInDeferredHydrationBoundary,
-  locateHydrationBoundaryClose,
-  locateHydrationNode,
-  markHydrationAnchor,
-  nextLogicalSibling,
-  setCurrentHydrationNode,
-} from './dom/hydration'
-import { isArray } from '@vue/shared'
+import { isHydrating, locateHydrationNode } from './dom/hydration'
 import { currentSlotOwner, setCurrentSlotOwner } from './componentSlots'
 import {
   type SlotBoundaryContext,
@@ -51,8 +31,9 @@ import {
   withOwnedSlotBoundary,
 } from './slotBoundary'
 import {
-  getCurrentSlotEndAnchor,
+  hydrateDynamicFragmentAnchor,
   isHydratingSlotFallbackActive,
+  prepareDeferredHydrationAnchor,
   setCurrentHydratingSlotFallbackActive,
   withHydratingSlotBoundary,
 } from './dom/hydrateFragment'
@@ -212,72 +193,13 @@ export class ForBlock extends VaporFragment {
   }
 }
 
-const enum CloseAnchorOwner {
-  None,
-  Self,
-  ParentBefore,
-  ParentAfter,
-}
-
-function getDynamicCloseOwner(
-  isSlot: boolean,
-  forwardedSlot: boolean,
-  anchorLabel: string | undefined,
-  nodes: Block,
-  currentSlotEndAnchor: Node | null,
-): CloseAnchorOwner {
-  // Slot fragments own the close marker unless this is an empty forwarded slot.
-  // Empty forwarded slots must leave the close marker to the parent boundary
-  // and create their runtime anchor after it.
-  if (isSlot) {
-    if (!forwardedSlot) return CloseAnchorOwner.Self
-    return isValidBlock(nodes)
-      ? CloseAnchorOwner.Self
-      : CloseAnchorOwner.ParentAfter
-  }
-
-  // SSR wraps multi-root `v-if` branches in a fragment range, so the closing
-  // `<!--]-->` belongs to the branch itself.
-  if (anchorLabel === 'if' && isArray(nodes) && nodes.length > 1) {
-    return CloseAnchorOwner.Self
-  }
-
-  // Slot fallback can fall through an inner `v-if`. When the `if` resolves
-  // to an invalid block and the fallback is selected, the `if` still needs
-  // its own runtime anchor instead of reusing the parent slot's end anchor.
-  if (
-    anchorLabel === 'if' &&
-    currentSlotEndAnchor &&
-    isHydratingSlotFallbackActive() &&
-    !isValidBlock(nodes)
-  ) {
-    return CloseAnchorOwner.ParentBefore
-  }
-
-  return CloseAnchorOwner.None
-}
-
-function queueAnchorInsert(
-  parentNode: Node,
-  nextNode: Node | null,
-  createAnchor: () => Node,
-): void {
-  // Create the runtime anchor only after insertion is flushed so traversal
-  // cannot observe a detached anchor too early.
-  queuePostFlushCb(() => {
-    const anchor =
-      nextNode && getParentNode(nextNode) === parentNode ? nextNode : null
-    parentNode.insertBefore(createAnchor(), anchor)
-  })
-}
-
 export class DynamicFragment extends VaporFragment {
   /**
    * @internal marker for duck typing to avoid direct instanceof check
    * which prevents tree-shaking of DynamicFragment
    */
   readonly __df = true
-  // @ts-expect-error - assigned in hydrate()
+  // @ts-expect-error - assigned in the constructor or hydrateDynamicFragmentAnchor()
   anchor: Node
   scope: EffectScope | undefined
   current?: BlockFn
@@ -323,7 +245,7 @@ export class DynamicFragment extends VaporFragment {
     if (key === this.current) {
       // On initial hydration, `key === current` means `render` is empty,
       // so this fragment hydrates as empty content.
-      if (isHydrating && !this.isSlot) this.hydrate(true)
+      if (isHydrating && !this.isSlot) hydrateDynamicFragmentAnchor(this, true)
       return
     }
 
@@ -418,34 +340,9 @@ export class DynamicFragment extends VaporFragment {
       }
     }
 
-    let reusingDeferredAnchor = false
-    if (isHydrating) {
-      const isRevivingDeferredBranch =
-        isInDeferredHydrationBoundary() &&
-        !!render &&
-        !this.isSlot &&
-        !isValidBlock(this.nodes)
-
-      reusingDeferredAnchor =
-        isRevivingDeferredBranch && !!this.anchor && !!this.anchor.parentNode
-
-      // Deferred hydration can keep an empty wrapper fragment alive, then resolve
-      // it to a real branch before hydration exits. Re-point the cursor at the
-      // fragment-owned insertion anchor so the late branch inserts before that
-      // anchor instead of consuming trailing hydrated siblings or the enclosing
-      // slot boundary.
-      if (isRevivingDeferredBranch) {
-        let slotEndAnchor: Node | null = null
-        const anchor =
-          this.anchor ||
-          (currentHydrationNode === (slotEndAnchor = getCurrentSlotEndAnchor())
-            ? slotEndAnchor
-            : null)
-        if (anchor) {
-          setCurrentHydrationNode(markHydrationAnchor(anchor))
-        }
-      }
-    }
+    const reusingDeferredAnchor = isHydrating
+      ? prepareDeferredHydrationAnchor(this, !!render)
+      : false
 
     this.renderBranch(
       render,
@@ -458,7 +355,7 @@ export class DynamicFragment extends VaporFragment {
     setActiveSub(prevSub)
 
     if (isHydrating && !this.isSlot && !reusingDeferredAnchor) {
-      this.hydrate(render == null)
+      hydrateDynamicFragmentAnchor(this, render == null)
     }
   }
 
@@ -543,247 +440,6 @@ export class DynamicFragment extends VaporFragment {
       onUpdated.forEach(hook => hook(this.nodes))
     }
   }
-
-  // Keep this as a prototype method to avoid per-instance closure allocation.
-  hydrate(isEmpty = false): void {
-    // early return allows tree-shaking of hydration logic when not used
-    if (!isHydrating) return
-
-    let advanceAfterRestore: Node | null = null
-    let exitHydrationBoundary: (() => void) | undefined
-
-    const reuseAnchor = (anchor: Node): void => {
-      this.anchor = markHydrationAnchor(anchor)
-      if (currentHydrationNode === this.anchor) {
-        advanceHydrationNode(this.anchor)
-      } else {
-        exitHydrationBoundary = enterHydrationBoundary(this.anchor)
-        advanceAfterRestore = this.anchor
-      }
-    }
-
-    const createRuntimeAnchor = (): Node =>
-      (this.anchor = markHydrationAnchor(
-        __DEV__ ? createComment(this.anchorLabel!) : createTextNode(),
-      ))
-
-    const cleanupAndInsertRuntimeAnchor = (
-      parentNode: Node,
-      nextNode: Node | null,
-      cleanupStart: Node,
-      cleanupUntil: Node | null,
-    ): void => {
-      if (cleanupUntil) {
-        exitHydrationBoundary = enterHydrationBoundary(cleanupUntil)
-      } else {
-        cleanupHydrationTail(cleanupStart)
-        setCurrentHydrationNode(null)
-      }
-      queueAnchorInsert(parentNode, nextNode, createRuntimeAnchor)
-    }
-
-    try {
-      // reuse `<!---->` as anchor
-      // `<div v-if="false"></div>` -> `<!---->`
-      if (isEmpty) {
-        if (isComment(currentHydrationNode!, '')) {
-          reuseAnchor(currentHydrationNode!)
-          return
-        }
-        if (
-          this.anchorLabel &&
-          currentHydrationNode &&
-          isComment(currentHydrationNode, 'teleport anchor')
-        ) {
-          const parentNode = getParentNode(currentHydrationNode)
-          const anchor = markHydrationAnchor(currentHydrationNode)
-          if (parentNode) {
-            // Target-side teleport anchors are structural. Empty dynamic
-            // fragments insert their own anchor before the target anchor
-            // instead of consuming it as mismatched SSR content.
-            queueAnchorInsert(parentNode, anchor, createRuntimeAnchor)
-            return
-          }
-        }
-        if (
-          !this.isSlot &&
-          this.anchorLabel &&
-          currentHydrationNode &&
-          !isHydratingSlotFallbackActive() &&
-          !isComment(currentHydrationNode, ']')
-        ) {
-          const parentNode = getParentNode(currentHydrationNode)
-          const anchor = nextLogicalSibling(currentHydrationNode)
-          // Empty branch against non-empty SSR output has no block node to
-          // derive an insertion point from, so use the current hydration range.
-          const reusableAnchor =
-            anchor &&
-            anchor.nodeType === 8 &&
-            isReusableDynamicFragmentAnchor(
-              anchor as Comment,
-              this.anchorLabel,
-            ) &&
-            getParentNode(anchor)
-              ? anchor
-              : null
-          if (parentNode) {
-            this.nodes = EMPTY_BLOCK
-            if (reusableAnchor) {
-              reuseAnchor(reusableAnchor)
-            } else {
-              cleanupAndInsertRuntimeAnchor(
-                parentNode,
-                anchor,
-                currentHydrationNode,
-                anchor,
-              )
-            }
-            return
-          }
-        }
-      }
-
-      // Reuse an existing SSR comment anchor for empty dynamic-component /
-      // async-component / keyed-fragment branches. Without this, hydration can
-      // end up creating a detached runtime anchor and lose the parent/sibling
-      // position needed for same-hydration branch flips.
-      if (
-        this.anchorLabel &&
-        !isValidBlock(this.nodes) &&
-        this.nodes instanceof Comment &&
-        isReusableDynamicFragmentAnchor(this.nodes, this.anchorLabel) &&
-        getParentNode(this.nodes)
-      ) {
-        const anchor = this.nodes
-        this.nodes = EMPTY_BLOCK
-        reuseAnchor(anchor)
-        return
-      }
-
-      // Empty dynamic fragments can also start from a detached runtime comment
-      // (for example client null against non-empty SSR content). In that case
-      // derive the insertion point from the current hydration cursor rather
-      // than from the detached block node, and let boundary cleanup trim the
-      // SSR range before the next logical sibling.
-      if (
-        this.anchorLabel &&
-        !isValidBlock(this.nodes) &&
-        this.nodes instanceof Comment &&
-        !getParentNode(this.nodes) &&
-        currentHydrationNode
-      ) {
-        const parentNode = getParentNode(currentHydrationNode)
-        const nextNode = nextLogicalSibling(currentHydrationNode)
-        if (parentNode) {
-          this.nodes = EMPTY_BLOCK
-          cleanupAndInsertRuntimeAnchor(
-            parentNode,
-            nextNode,
-            currentHydrationNode,
-            nextNode,
-          )
-          return
-        }
-      }
-
-      const currentSlotEndAnchor = getCurrentSlotEndAnchor()
-      const forwardedSlot = this.isSlot
-        ? (this as any as SlotFragment).forwarded
-        : false
-      const slotAnchor = this.isSlot ? currentSlotEndAnchor : null
-
-      // Reuse SSR `<!--]-->` as anchor.
-      // SSR wraps slots and multi-root `v-if` branches with `<!--[-->...<!--]-->`.
-      // Non-forwarded slots always own the closing `<!--]-->`, even when empty.
-      // Forwarded slots only own it when they rendered valid content.
-      const closeOwner = getDynamicCloseOwner(
-        !!this.isSlot,
-        forwardedSlot,
-        this.anchorLabel,
-        this.nodes,
-        currentSlotEndAnchor,
-      )
-      if (closeOwner === CloseAnchorOwner.Self) {
-        const anchor = locateHydrationBoundaryClose(
-          slotAnchor || currentHydrationNode!,
-          slotAnchor || null,
-        )
-        if (isComment(anchor!, ']')) {
-          reuseAnchor(anchor)
-          return
-        } else if (__DEV__) {
-          throw new Error(
-            `Failed to locate ${this.anchorLabel} fragment anchor. this is likely a Vue internal bug.`,
-          )
-        }
-      } else if (
-        closeOwner === CloseAnchorOwner.ParentAfter &&
-        currentSlotEndAnchor
-      ) {
-        // Otherwise, create a new anchor.
-        // This covers: empty forwarded slots.
-        // Keep the forwarded slot close marker structural for parent cleanup,
-        // even though this fragment uses a runtime anchor after it.
-        const anchor = markHydrationAnchor(currentSlotEndAnchor)
-        queueAnchorInsert(
-          anchor.parentNode!,
-          anchor.nextSibling,
-          createRuntimeAnchor,
-        )
-        return
-      } else if (
-        closeOwner === CloseAnchorOwner.ParentBefore &&
-        currentSlotEndAnchor
-      ) {
-        const endAnchor = currentSlotEndAnchor
-        queuePostFlushCb(() => {
-          const parentNode = getParentNode(endAnchor)
-          if (!parentNode) return
-          parentNode.insertBefore(createRuntimeAnchor(), endAnchor)
-        })
-        return
-      }
-
-      // Otherwise, create a new anchor.
-      // This covers: dynamic-component, async component, keyed fragment.
-      let parentNode: Node | null
-      let nextNode: Node | null
-      if (
-        this.anchorLabel === 'if' &&
-        !isValidBlock(this.nodes) &&
-        currentSlotEndAnchor &&
-        currentHydrationNode === currentSlotEndAnchor
-      ) {
-        // Only reuse the slot end anchor as insertion point when this empty
-        // inner `v-if` has already consumed the whole local slot range.
-        parentNode = currentSlotEndAnchor.parentNode
-        nextNode = currentSlotEndAnchor
-      } else {
-        const node = findBlockBoundary(this.nodes)
-        parentNode = node.parentNode
-        nextNode = node.nextNode
-      }
-      queueAnchorInsert(parentNode!, nextNode, createRuntimeAnchor)
-    } finally {
-      exitHydrationBoundary && exitHydrationBoundary()
-      if (advanceAfterRestore && currentHydrationNode === advanceAfterRestore) {
-        advanceHydrationNode(advanceAfterRestore)
-      }
-    }
-  }
-}
-
-function isReusableDynamicFragmentAnchor(
-  node: Comment,
-  anchorLabel: string,
-): boolean {
-  return (
-    isComment(node, anchorLabel) ||
-    (isComment(node, '') &&
-      (anchorLabel === 'dynamic-component' ||
-        anchorLabel === 'async component' ||
-        anchorLabel === 'keyed'))
-  )
 }
 
 // SlotFragment must live in the same module as DynamicFragment: `extends`
@@ -898,12 +554,13 @@ export class SlotFragment extends DynamicFragment implements SlotFallbackState {
             // Updates run under the temporary fallback-active marker so empty
             // inner branches can materialize their own anchors if fallback
             // takes over. If recheck resolves back to content, restore the
-            // outer state before hydrate(); the surrounding finally still
-            // restores nested callers when we leave this boundary.
+            // outer state before hydrateDynamicFragmentAnchor(); the surrounding
+            // finally still restores nested callers when we leave this
+            // boundary.
             if (!hasSlotFallback(boundary) || contentValid) {
               setCurrentHydratingSlotFallbackActive(prev)
             }
-            this.hydrate(!isValidBlock(this.nodes))
+            hydrateDynamicFragmentAnchor(this, !isValidBlock(this.nodes))
           } finally {
             setCurrentHydratingSlotFallbackActive(prev)
           }
